backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
numpy==1.24.3
opencv-python==4.9.0.80
tensorflow-cpu==2.15.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
